+++ config/routes.rb
@@ -5,14 +5,16 @@ Rails.application.routes.draw do
 
   get '/home', to: 'homes#index'
   get '/plants', to: 'homes#index'
-  # post '/plants', to: 'plants#create'
-
-  resources :plants, only: [:create]    #separate out because I don't need to make my own API???
 
   namespace :api do
     namespace :v1 do
-      post '/plants/search', to:'plants#search'
-      # resources :plants, only: [:create]
+      post '/spoonacular/search', to:'spoonacular#search'
+    end
+  end
+
+  namespace :api do
+    namespace :v1 do
+      resources :plants, only: [:create]
     end
   end
